+++ scipy/special/special/exprel.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include "config.h"
+
+namespace special {
+
+inline double exprel(double x) {
+    if (std::abs(x) < std::numeric_limits<double>::epsilon()) {
+        return 1;
+    }
+
+    if (x > 717) { // near log(DBL_MAX)
+        return std::numeric_limits<double>::infinity();
+    }
+
+    return std::expm1(x) / x;
+}
+
+inline float exprel(float x) { return exprel(static_cast<double>(x)); }
+
+} // namespace special
